+++ Dockerfile
@@ -4,7 +4,7 @@ EXPOSE 80
 WORKDIR /app
 
 RUN apt update && apt upgrade -y
-RUN apt install nodejs -y
+RUN apt install nodejs npm -y
 
 RUN apt install nginx -y
 
@@ -13,5 +13,7 @@ RUN touch ./docker
 COPY ["./assets","./assets"]
 COPY ["./src","./src"]
 COPY ["./package.json","./package.json"]
+
+RUN npm install
 
 ENTRYPOINT ["node", "./src/main.js"]
